a function that computes the CBF Constraints
For each link, 2 main circles are placed at the start and end of it, and an additional number of circles are assigned by the user whose positions are distributed evenly along each link length, then the total circles are  4 + (circles/link)* links, which also gives the total number of constraints.  For ex, for 1 additional circle per link, the total constraints are 7. 
This file needs to be included in any path stored in your MATLAB directory before being able to use it in the main NMPC file to generate simulations.